Refactor: Renamed IFileService as FileService
Refactor: Renamed FileService as FileServiceImpl
