+++ src/main/java/com/medipol/h5210017/midterm/service/file/FileService.java
@@ -1,8 +1,8 @@
-package com.medipol.h5210017.midterm.service;
+package com.medipol.h5210017.midterm.service.file;
 
 import java.io.File;
 import java.io.IOException;
 
-public interface IFileService {
+public interface FileService {
     File create(String path, String content) throws IOException;
 }
